refactor(ai): address review on async stream wrapper
- Wrap the generator close in try/finally so the underlying provider
  stream is always closed even if the PostHog capture in the generator's
  finally raises (avoids leaking the HTTP connection).
- Extract the duplicated mock stream into a shared
  posthog/test/ai/utils.py::RecordingAsyncStream used by all three test
  modules.
- Trim verbose comments/docstrings.
